Merge together divergent code branches
It seems like at some point we ended up with divergent branches in
this repo which had slightly different changes applied to them,
including differing merges from the upstream. Our monolith currently
targets the commit `ee7a388`, but this commit has differing migration
history from what is actually in the app. Additionally, the default
branch `develop` has had a variety of security changes applied to it
(although not strictly necessary since we don't use the requirements
from this repo in any way).

To resolve this, I have taken the base `develop` branch and merged in
the commit `ee7a388`. I've manually picked changes that are relevant
and kept the correct migration history. The code in this newly merged
branch should contain the relevant code that we are currently running
via the monolith, but will give us a clean point to reference from now
moving forward.